rename(lotto, lotto-test): 패키지 변경
